fix(OMN-12909): satisfy gateway forwarder validators
