+++ tests/unit/nodes/node_bus_forwarder_effect/test_gateway_forwarder_service.py
@@ -24,6 +24,9 @@ from omnibase_infra.nodes.node_bus_forwarder_effect.services.service_gateway_for
 pytestmark = pytest.mark.asyncio
 
 TENANT_ID = UUID("11111111-1111-1111-1111-111111111111")
+BROKER_PROVIDER_ID = UUID("22222222-2222-2222-2222-222222222222")
+PRINCIPAL_ID = UUID("33333333-3333-3333-3333-333333333333")
+CORRELATION_ID = UUID("44444444-4444-4444-4444-444444444444")
 INBOUND_TOPIC = "onex.cmd.omnibase-infra.delegation-inference-request.v1"
 OUTBOUND_TOPIC = "onex.evt.omnibase-infra.inference-response.v1"
 WIRE_INBOUND_TOPIC = f"tenant-acme.{INBOUND_TOPIC}"
@@ -91,10 +94,10 @@ def _config() -> ModelGatewayForwarderConfig:
         tenant_identity=ModelGatewayTenantIdentity(
             tenant_id=TENANT_ID,
             tenant_slug="acme",
-            principal_id=f"tenant:{TENANT_ID}",
+            principal_id=PRINCIPAL_ID,
         ),
         cloud_bus=ModelGatewayCloudBusConfig(
-            broker_provider_id="redpanda-dogfood",
+            broker_provider_id=BROKER_PROVIDER_ID,
             cloud_broker_ref="gateway.cloud.kafka.broker",
             cloud_auth_ref="gateway.cloud.kafka.oauth",
             acl_provisioner_ref="gateway.cloud.kafka.authorization",
@@ -113,8 +116,8 @@ def _envelope(**overrides: object) -> ModelGatewayEnvelope:
     values = {
         "tenant_id": TENANT_ID,
         "tenant_slug": "acme",
-        "envelope_id": str(uuid4()),
-        "correlation_id": "corr-1",
+        "envelope_id": uuid4(),
+        "correlation_id": CORRELATION_ID,
         "causation_id": "cause-1",
         "event_type": "LlmInferenceResponse",
         "source_topic": OUTBOUND_TOPIC,
@@ -163,7 +166,7 @@ async def test_outbound_local_message_is_published_to_cloud_wire_topic() -> None
     forwarded = ModelGatewayEnvelope.model_validate_json(published.value)
     assert forwarded.wire_topic == WIRE_OUTBOUND_TOPIC
     assert forwarded.canonical_topic == OUTBOUND_TOPIC
-    assert forwarded.correlation_id == "corr-1"
+    assert forwarded.correlation_id == CORRELATION_ID
     assert forwarded.event_type == "LlmInferenceResponse"
 
 
@@ -193,7 +196,7 @@ async def test_inbound_cloud_message_is_published_to_local_canonical_topic() -> 
     forwarded = ModelGatewayEnvelope.model_validate_json(published.value)
     assert forwarded.source_topic == WIRE_INBOUND_TOPIC
     assert forwarded.canonical_topic == INBOUND_TOPIC
-    assert forwarded.correlation_id == "corr-1"
+    assert forwarded.correlation_id == CORRELATION_ID
     assert forwarded.event_type == "DelegationInferenceRequest"
