+++ tests/unit/nodes/node_bus_forwarder_effect/test_gateway_topic_transform.py
@@ -20,6 +20,9 @@ from omnibase_infra.nodes.node_bus_forwarder_effect.models import (
 )
 
 TENANT_ID = UUID("11111111-1111-1111-1111-111111111111")
+BROKER_PROVIDER_ID = UUID("22222222-2222-2222-2222-222222222222")
+PRINCIPAL_ID = UUID("33333333-3333-3333-3333-333333333333")
+CORRELATION_ID = UUID("44444444-4444-4444-4444-444444444444")
 INBOUND_TOPIC = "onex.cmd.omnibase-infra.delegation-inference-request.v1"
 OUTBOUND_TOPIC = "onex.evt.omnibase-infra.inference-response.v1"
 
@@ -29,10 +32,10 @@ def _config() -> ModelGatewayForwarderConfig:
         tenant_identity=ModelGatewayTenantIdentity(
             tenant_id=TENANT_ID,
             tenant_slug="acme",
-            principal_id=f"tenant:{TENANT_ID}",
+            principal_id=PRINCIPAL_ID,
         ),
         cloud_bus=ModelGatewayCloudBusConfig(
-            broker_provider_id="redpanda-dogfood",
+            broker_provider_id=BROKER_PROVIDER_ID,
             cloud_broker_ref="gateway.cloud.kafka.broker",
             cloud_auth_ref="gateway.cloud.kafka.oauth",
             acl_provisioner_ref="gateway.cloud.kafka.authorization",
@@ -51,8 +54,8 @@ def _envelope(**overrides: object) -> ModelGatewayEnvelope:
     values = {
         "tenant_id": TENANT_ID,
         "tenant_slug": "acme",
-        "envelope_id": str(uuid4()),
-        "correlation_id": "corr-1",
+        "envelope_id": uuid4(),
+        "correlation_id": CORRELATION_ID,
         "causation_id": None,
         "event_type": "LlmInferenceResponse",
         "source_topic": OUTBOUND_TOPIC,
@@ -67,7 +70,7 @@ def _envelope(**overrides: object) -> ModelGatewayEnvelope:
 def test_outbound_adds_tenant_wire_prefix_and_preserves_event_fields() -> None:
     envelope = _envelope()
 
-    result = HandlerForwardOutbound(_config()).handle(envelope)
+    result = HandlerForwardOutbound(_config()).forward_outbound(envelope)
 
     assert result.wire_topic == f"tenant-acme.{OUTBOUND_TOPIC}"
     assert result.canonical_topic == OUTBOUND_TOPIC
@@ -79,7 +82,7 @@ def test_outbound_rejects_undeclared_topic() -> None:
     envelope = _envelope(canonical_topic="onex.evt.omnibase-infra.not-declared.v1")
 
     with pytest.raises(ValueError, match="not declared"):
-        HandlerForwardOutbound(_config()).handle(envelope)
+        HandlerForwardOutbound(_config()).forward_outbound(envelope)
 
 
 def test_inbound_strips_tenant_prefix_to_bare_contract_topic() -> None:
@@ -90,7 +93,7 @@ def test_inbound_strips_tenant_prefix_to_bare_contract_topic() -> None:
         canonical_topic=INBOUND_TOPIC,
     )
 
-    result = HandlerConsumeInbound(_config()).handle(envelope)
+    result = HandlerConsumeInbound(_config()).consume_inbound(envelope)
 
     assert result.source_topic == f"tenant-acme.{INBOUND_TOPIC}"
     assert result.canonical_topic == INBOUND_TOPIC
@@ -106,7 +109,7 @@ def test_inbound_rejects_cross_tenant_wire_prefix() -> None:
     )
 
     with pytest.raises(ValueError, match="tenant prefix"):
-        HandlerConsumeInbound(_config()).handle(envelope)
+        HandlerConsumeInbound(_config()).consume_inbound(envelope)
 
 
 def test_inbound_rejects_envelope_tenant_mismatch() -> None:
@@ -118,4 +121,4 @@ def test_inbound_rejects_envelope_tenant_mismatch() -> None:
     )
 
     with pytest.raises(ValueError, match="tenant_id"):
-        HandlerConsumeInbound(_config()).handle(envelope)
+        HandlerConsumeInbound(_config()).consume_inbound(envelope)
